--- web/backend/service/flask_app/sc_memory.py
@@ -1,4 +1,0 @@
-from .settings import settings
-from sc.memory import Memory
-
-memory = Memory(settings.get_path_to_sc_config())
