--- web/backend/service/flask_app/settings.py
@@ -1,4 +1,0 @@
-from ..config import Config
-import os
-
-settings = Config(os.environ.get("CONFIG_PATH"))
