+++ web/backend/service/sc_memory.py
@@ -1,0 +1,4 @@
+from sc.memory import Memory
+from .config import config
+
+memory = Memory(config.get_path_to_sc_config())
